+++ src/main/java/com/openprise/dao/ParticipantRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.openprise.domain.Participant;
 
-public interface ParticipantRepository  extends CrudRepository<Participant, Long> {
+public interface ParticipantRepository  extends CrudRepository<Participant, Integer> {
 
 	Participant findByMobileAndPassword(String mobile, String password);
 	Participant findByMobile(String mobile);
