+++ src/main/java/com/openprise/dao/ParticipantTestRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.openprise.domain.Test;
 import com.openprise.domain.TestParticipant;
 
-public interface ParticipantTestRepository  extends CrudRepository<TestParticipant, Long> {
+public interface ParticipantTestRepository  extends CrudRepository<TestParticipant, Integer> {
 	
 	Collection<TestParticipant> findByTest(Test test);
