+++ src/main/java/com/openprise/dao/QuestionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.openprise.domain.Question;
 
-public interface QuestionRepository  extends CrudRepository<Question, Long> {
+public interface QuestionRepository  extends CrudRepository<Question, Integer> {
 
 }
