+++ src/main/java/com/openprise/dao/TestRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.openprise.domain.Test;
 
-public interface TestRepository  extends CrudRepository<Test, Long> {
+public interface TestRepository  extends CrudRepository<Test, Integer> {
 
 }
